cmd/bench: Move statistics-related functions into separate package
