+++ cmd/bench/stats/stat-interceptor.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package cmd
+package stats
 
 import (
 	"github.com/filecoin-project/mir/pkg/events"
